Created a new cache engine for redis cluster. Move common functionality into a trait and wrap the few operations that work differently.
