+++ src/Cache/Cache.php
@@ -83,6 +83,7 @@ class Cache
         'memcached' => Engine\MemcachedEngine::class,
         'null' => Engine\NullEngine::class,
         'redis' => Engine\RedisEngine::class,
+        'rediscluster' => Engine\RedisClusterEngine::class,
     ];
 
     /**
